feat: CardRegister 로직 분리
